requirements.txt
azure-kusto-data
numpy
pandas
seaborn
python-dotenv